+++ commands/infos.js
@@ -1,39 +1,33 @@
 module.exports = {
     name: "infos",
-    alias: ["informations", "ping"],
-    description: "Get the bot's information",
+    alias: ["informations", "getinfos"],
+    description: "Get informations about a server, a channel, a role or a member",
     ownersOnly: false,
     active: true,
-    type: "informations",
-    color: "#13D7A3",
+    type: "usefull",
+    color: "#00b0b3",
+    arguments: "<server> | <channel <[channel_mention] | [channel_id] | [channel_name]>> | <role [role_mention] | [role_name] | [role_id]>",
     deleteCommand: false,
     permissions: {
         bot: [],
         user: []
     },
     run: (e, vars, args, settings) => {
-        let informations = {
-            version: vars.package.version,
-            ping: vars.client.ws.ping + 'ms',
-            "invite link": `[click here to add me](https://discord.com/api/oauth2/authorize?client_id=${vars.client.user.id}&permissions=8&scope=bot)`,
-            "discord support server": `[click here to join it](https://discord.gg/B6cGv6hyHR)`,
-            "top.gg page": `[click here to view ${vars.client.user.username}'s top.gg page](https://top.gg/bot/${vars.client.user.id})`
+        let type = args.shift()
+        if(!type) return e.reply('You need to indicate a type (server, channel, role or user)')
+
+        try {
+            var command = require('./infos/' + type)
+        } catch (err) {
+            vars.log(err)
+            return e.reply('Invalid command argument!')
         }
 
-        for(let owner of vars.configs.owners) {
-            if(!informations.author) informations.author = ""
-            if(informations.author) informations.author += " & "
-            informations.author += vars.client.users.cache.get(owner).tag
+        try {
+            command.run(e, vars, args, settings)
+        } catch (err) {
+            vars.log(err)
+            return e.reply('An error has occured...')
         }
-
-        let embed = new vars.discord.MessageEmbed()
-        .setColor(this.color || "RANDOM")
-        .setAuthor("This is my informations:", vars.client.user.avatarURL())
-        .setThumbnail(vars.assets.images.informations)
-        for(let info in informations) {
-            embed.addField("__" + info + ":__", "> " + informations[info])
-        }
-
-        e.channel.send(embed)
     }
 }