+++ commands/settings/set.js
@@ -1,3 +1,4 @@
+const { version } = require('discord.js')
 const {setSettings} = require('../settings')
 module.exports = {
     name: "set",
@@ -7,7 +8,7 @@ module.exports = {
         if(!args[0]){
             var channel = e.channel
         }else {
-            var channel = e.mentions.channels.first() || vars.client.guilds.cache.get(e.guild.id).channels.cache.get(args[0]) || vars.client.guilds.cache.get(e.guild.id).channels.cache.find(c => c.name.toLowerCase() === args[0].toLowerCase() && c.isText())
+            var channel = e.mentions.channels.first() || e.guild.channels.cache.get(args[0]) || e.guild.channels.cache.find(c => c.name.toLowerCase() === args[0].toLowerCase() && c.isText())
         }
 
         if(!channel || !channel.isText()) return e.reply('Please indicate a text channel!')